feat: added integration tests
