+++ out.sh
@@ -1,2 +1,4 @@
 a=12;
-a=12
+b=2;
+a=$(echo $a '+' 12 | bc -l | sed '/\./ s/\.\{0,1\}0\{1,\}$//');
+echo ($(echo $a '+' $b | bc -l | sed '/\./ s/\.\{0,1\}0\{1,\}$//'))
