+++ test.amber
@@ -1,3 +1,4 @@
 let a = 12
-a = a + 12
-$echo {a}$
+let b = 2
+let c = 3
+$echo {(a + b) * c}$
